Added duplication of local setup to remote sites through eups setup. This is set up here and available to the ctrl_platform_* templates, which can use this new keyword (CTRL_EXECUTE_SETUP_PACKAGES), or not, as it chooses.  This variable can be used to substitute a multi-line block of setups, that look something like this:
  setup -j foo 1.0
  setup -j bar 2.2
  ...

Locally setup packages are NOT included in this block, and are
specifically ignored, since the local file paths mean nothing on the
platforms that would need this.  The local cluster, which has a shared
file space and homogeneous operating environments on all the machines
uses the condor "getenv" method of transfering this information, so for
ctrl_platform_lsst this is not used.