Added --setup option to set up different package names on a remote system. (This won't work locally, because we transfer the whole environment over via the Condor "getenv" command).
Added lookup user name and home in $HOME/.lsst/condor-info.py.
If running on the lsst cluster, lookup the names in that Config first,
and if they're not there, default to the login name and $HOME,
respectively.

Use setup -j to write out values of setup products.  The inline
python script takes care of extracting these from the remote environment
and writing them to env.sh.

Bug fix:  Local scratch is in terms of local home directory, not remote
home directory.