Changed configurations to use DatabaseConfigurator stand-in plug-in
The DatabaseConfigurator is a base class that is a placeholder for
a "real" implementation of a database plugin.  Since at this time we
don't need to have any interactions with the database (because a
replacement for provenance hasn't been implemented), this information
is unused.  It could be deleted without causing errors in orchestration,
but is being kept here because it will be used in the future.